chore: Cleaned up the packages by removing and moving around java classes; Also added a server port property in order to map in the future other webservers appropriately
