--- src/main/userContext/src/main/java/com/userContext/application_layer/ApplicationAPI.java
@@ -1,10 +1,0 @@
-package com.userContext.application_layer;
-
-import com.userContext.application_layer.exceptions.UserAlreadyExistsException;
-
-public interface ApplicationAPI {
-
-    void registerUser(final String id, final String name, final String surname, final String role) throws UserAlreadyExistsException;
-
-
-}
