+++ src/main/userContext/src/main/java/com/userContext/application_layer/UserNotFoundAdvice.java
@@ -1,6 +1,6 @@
-package com.userContext.infrastructure_layer.springBoot;
+package com.userContext.application_layer;
 
-import com.userContext.infrastructure_layer.springBoot.exceptions.UserNotFoundException;
+import com.userContext.application_layer.exceptions.UserNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -8,12 +8,12 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ControllerAdvice
-public class UserNotFoundAdvice {
+class UserNotFoundAdvice {
 
     @ResponseBody
     @ExceptionHandler(UserNotFoundException.class)
     @ResponseStatus(HttpStatus.NOT_FOUND)
-    String employeeNotFoundHandler(UserNotFoundException ex) {
+    String userNotFoundHandler(UserNotFoundException ex) {
         return ex.getMessage();
     }
