--- src/main/userContext/src/main/java/com/userContext/application_layer/exceptions/UserAlreadyExistsException.java
@@ -1,5 +1,0 @@
-package com.userContext.application_layer.exceptions;
-
-public class UserAlreadyExistsException extends Exception {
-    public UserAlreadyExistsException(){}
-}
