+++ src/main/userContext/src/main/java/com/userContext/application_layer/exceptions/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.userContext.application_layer.exceptions;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(Long id) {
+        super("Could not find user " + id);
+    }
+}
