--- src/main/userContext/src/main/java/com/userContext/business_logic_layer/DataSourcePort.java
@@ -1,9 +1,0 @@
-package com.userContext.business_logic_layer;
-
-import com.userContext.business_logic_layer.exceptions.DataSourceException;
-
-public interface DataSourcePort {
-
-    void saveUser() throws DataSourceException;
-
-}
