+++ src/main/userContext/src/main/java/com/userContext/business_logic_layer/User.java
@@ -25,18 +25,7 @@ public class User implements Serializable {
         this.role = role;
     }
 
-    public User() {
-
-    }
-
-    public void saveUser() {
-        try {
-            DomainModelImpl.getDataSourcePort().saveUser();
-        } catch (Exception ex){
-            ex.printStackTrace();
-        }
-
-    }
+    public User() {}
 
     public Long getId() {
         return id;
