+++ src/main/userContext/src/main/java/com/userContext/business_logic_layer/UserRepository.java
@@ -1,4 +1,4 @@
-package com.userContext.infrastructure_layer.springBoot;
+package com.userContext.business_logic_layer;
 
 import com.userContext.business_logic_layer.User;
 import org.springframework.data.jpa.repository.JpaRepository;
