--- src/main/userContext/src/main/java/com/userContext/business_logic_layer/exceptions/DataSourceException.java
@@ -1,5 +1,0 @@
-package com.userContext.business_logic_layer.exceptions;
-
-public class DataSourceException extends Exception {
-    public DataSourceException(final String msg) {super(msg);}
-}
