--- src/main/userContext/src/main/java/com/userContext/infrastructure_layer/springBoot/exceptions/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.userContext.infrastructure_layer.springBoot.exceptions;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(Long id) {
-        super("Could not find employee " + id);
-    }
-}
