+++ src/main/userContext/src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.datasource.driver-class-name=org.h2.Driver
 
 # jpa/hibernate
 spring.jpa.show-sql = false
-spring.jpa.hibernate.ddl-auto = create
+spring.jpa.hibernate.ddl-auto = none
 spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
 spring.jpa.database-platform = org.hibernate.dialect.H2Dialect
